cloud-db/ai_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
requests>=2.32.0
pydantic>=2.0.0
openai>=1.60.0
google-auth>=2.29.0
python-jose[cryptography]>=3.3.0
psycopg[binary]>=3.1.18
